DP-GA program
For the 2D-COF structure, we selected specific linkers, nodes and topological structures, and used grand canonical Monte Carlo simulation to screen 2D-COF materials with excellent performance. We developed a program combining deep learning and genetic algorithms to predict the excellent structures of each generation, perform accurate GCMC calculations on the newly generated excellent structures that are not in the database and add them to the database for the next iterative prediction, ultimately obtaining a better structure.